requirements.txt
Pillow
numpy
termcolor
